requirements.txt
dj-database-url==0.5.0
Django==2.2.13
django-heroku==0.3.1
gunicorn==19.9.0
psycopg2==2.8.3
pytz==2019.1
sqlparse==0.3.0
whitenoise==4.1.3
djangorestframework==3.11.0
requests==2.23.0

